Implement throwErrorOnAssetMissing configuration
We use javascript hot-reloading in development which means the asset
doesn't exist in the file system but is generated on the file by a node
server.

This configuration param will determine if an error is thrown on missing
assets.  By default, it will throw an exception. If set to false, a
cache busting string is still included using a randomString method.

Fixes #9